requirements.txt
fastapi
uvicorn
python-dotenv
openai
azure-ai-projects
azure-identity
pandas
sqlalchemy
pyodbc